Razor Fun!
Create a new string array in a View, using Razor.  Then, iterate through the array - generating a <p> tag for each iteration, and placing the contents on the string from your array within the inner text of the <p> element.
BONUS:  If the string in the array begins with the letter c, style the <p> element with red text!